test(cucumber): isolate worker delay fixture
